Begin moving from Soya3D to Panda3D
Rationale is that Panda3D has much better support and a much better
community. Eventually we need to encapsulate Panda3D calls underneath
custom modules so that if we ever need to change libraries (again) in
the future, it won't be so painful. For now getting it working is a
priority. With this commit, a simple scene with an oven and grass
ground is visible.