Integrate AI module
Since Panda3D's AI is (logically) heavily integrated with the rest of the
Panda3D system, I'm putting it under the graphics module so that game data
does not need to make a call to both graphics and ai every time a move
command is given. Right now we just have a buggy AI that simply moves
towards the point specified by the user.